requirements.txt
Flask==2.3.3
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
validators==0.22.0
whois-python==0.4.3
dnspython==2.4.2
urllib3==2.0.7
lxml==4.9.3
flask-cors==4.0.0 